User can now register for a full account with a password. Password is now required to login

+++ src/components/auth/Register.js
@@ -5,7 +5,8 @@ import "./Login.css"
 
 export const Register = () => {
 
-    const [registerUser, setRegisterUser] = useState({ firstName: "", lastName: "", email: "" })
+    const [registerUser, setRegisterUser] = useState({ firstName: "", lastName: "", email: "", password: "", trackValue: 0.0104 })
+    const [passwordConfirm, setPasswordConfirm] = useState("")
     const [conflictDialog, setConflictDialog] = useState(false)
 
     const history = useHistory()
@@ -17,7 +18,7 @@ export const Register = () => {
     }
 
     const existingUserCheck = () => {
-        
+
         return fetch(`${authApi.localApiBaseUrl}/${authApi.endpoint}?email=${registerUser.email}`)
             .then(res => res.json())
             .then(user => !!user.length)
@@ -29,23 +30,30 @@ export const Register = () => {
         existingUserCheck()
             .then((userExists) => {
                 if (!userExists) {
-                    fetch(`${authApi.localApiBaseUrl}/${authApi.endpoint}`, {
-                        method: "POST",
-                        headers: {
-                            "Content-Type": "application/json"
-                        },
-                        body: JSON.stringify({
-                            email: registerUser.email,
-                            name: `${registerUser.firstName} ${registerUser.lastName}`
+                    if (registerUser.password === passwordConfirm) {
+                        fetch(`${authApi.localApiBaseUrl}/${authApi.endpoint}`, {
+                            method: "POST",
+                            headers: {
+                                "Content-Type": "application/json"
+                            },
+                            body: JSON.stringify({
+                                email: registerUser.email,
+                                firstName: registerUser.firstName,
+                                lastName: registerUser.lastName,
+                                password: registerUser.password,
+                                userTrackValue: parseFloat(registerUser.trackValue)
+                            })
                         })
-                    })
-                        .then(res => res.json())
-                        .then(createdUser => {
-                            if (createdUser.hasOwnProperty("id")) {
-                                sessionStorage.setItem(userStorageKey, createdUser.id)
-                                history.push("/")
-                            }
-                        })
+                            .then(res => res.json())
+                            .then(createdUser => {
+                                if (createdUser.hasOwnProperty("id")) {
+                                    sessionStorage.setItem(userStorageKey, createdUser.id)
+                                    history.push("/")
+                                }
+                            })
+                    } else {
+                        window.alert('Passwords do not match')
+                    }
                 }
                 else {
                     setConflictDialog(true)
@@ -63,7 +71,8 @@ export const Register = () => {
             </dialog>
 
             <form className="form--login" onSubmit={handleRegister}>
-                <h1 className="h3 mb-3 font-weight-normal">Please Register for Application Name</h1>
+                <h1 className="h3 mb-3 font-weight-normal">Howdy, stranger!</h1>
+                <h2 className="h3 mb-3 font-weight-normal">Tell us about yourself</h2>
                 <fieldset>
                     <label htmlFor="firstName"> First Name </label>
                     <input type="text" name="firstName" id="firstName" className="form-control" placeholder="First name" required autoFocus value={registerUser.firstName} onChange={handleInputChange} />
@@ -75,6 +84,20 @@ export const Register = () => {
                 <fieldset>
                     <label htmlFor="inputEmail"> Email address </label>
                     <input type="email" name="email" id="email" className="form-control" placeholder="Email address" required value={registerUser.email} onChange={handleInputChange} />
+                </fieldset>
+                <fieldset>
+                    <label htmlFor="password"> Password </label>
+                    <input type="password" name="password" id="password" className="form-control" placeholder="password" required value={registerUser.password} onChange={handleInputChange} />
+                </fieldset>
+                <fieldset>
+                    <label htmlFor="confirm"> Confirm Password </label>
+                    <input type="password" name="confirm" id="confirm" className="form-control" placeholder="confirm password" required onChange={e => setPasswordConfirm(e.target.value)} />
+                </fieldset>
+                <fieldset>
+                    <label htmlFor="trackValue"> Track Value: $</label>
+                    <input type="number" step="0.0001" name="trackValue" id="trackValue" className="form-control" defaultValue="0.0104" required onChange={handleInputChange} />
+                    <p><strong>What is this?</strong></p>
+                    <p>You may set a default value per track for your reports. If you're not sure, leave it as the default value ($0.0104) for now, and we'll talk about it later. You'll be able to change this later as well. </p>
                 </fieldset>
                 <fieldset>
                     <button type="submit"> Sign in </button>
